Revert "Rename receice.py to response.py."
This reverts commit 16413e720226b75e40ea0c48eaf69d09b0580277.
