+++ tests/unit/command_runners/lsp/receive_test.py
@@ -3,7 +3,7 @@ from asyncio import StreamReader
 
 import pytest
 
-from pre_commit_sbt.command_runners.lsp.response import read_until_complete_message
+from pre_commit_sbt.command_runners.lsp.receive import read_until_complete_message
 from pre_commit_sbt.err.error_msgs import COMMAND_FAILED
 from pre_commit_sbt.err.exceptions import FailedCommandError
